move getHiddenElements to otherActions.tsx
